requirements.txt
gunicorn==20.1.0
django==4.2
mysqlclient==2.2
Pillow==8.3.2
django-debug-toolbar==4.2.0
django-allauth
django-extensions
django-cleanup
django-sass-processor
geocoder
openai==0.28
awscli
boto3
matplotlib seaborn
faker
django_seed